format the project, delete unit tests

+++ decorate-angular-cli.js
@@ -21,24 +21,24 @@
  * - Delete and reinstall your node_modules
  */
 
-const fs = require("fs");
-const os = require("os");
-const cp = require("child_process");
-const isWindows = os.platform() === "win32";
-const { output } = require("@nrwl/workspace");
+const fs = require('fs');
+const os = require('os');
+const cp = require('child_process');
+const isWindows = os.platform() === 'win32';
+const { output } = require('@nrwl/workspace');
 
 /**
  * Paths to files being patched
  */
-const angularCLIInitPath = "node_modules/@angular/cli/lib/cli/index.js";
+const angularCLIInitPath = 'node_modules/@angular/cli/lib/cli/index.js';
 
 /**
  * Patch index.js to warn you if you invoke the undecorated Angular CLI.
  */
 function patchAngularCLI(initPath) {
-  const angularCLIInit = fs.readFileSync(initPath, "utf-8").toString();
+  const angularCLIInit = fs.readFileSync(initPath, 'utf-8').toString();
 
-  if (!angularCLIInit.includes("NX_CLI_SET")) {
+  if (!angularCLIInit.includes('NX_CLI_SET')) {
     fs.writeFileSync(
       initPath,
       `
@@ -47,7 +47,7 @@ if (!process.env['NX_CLI_SET']) {
   output.warn({ title: 'The Angular CLI was invoked instead of the Nx CLI. Use "npx ng [command]" or "nx [command]" instead.' });
 }
 ${angularCLIInit}
-    `
+    `,
     );
   }
 }
@@ -58,16 +58,15 @@ ${angularCLIInit}
  */
 function symlinkNgCLItoNxCLI() {
   try {
-    const ngPath = "./node_modules/.bin/ng";
-    const nxPath = "./node_modules/.bin/nx";
+    const ngPath = './node_modules/.bin/ng';
+    const nxPath = './node_modules/.bin/nx';
     if (isWindows) {
       /**
        * This is the most reliable way to create symlink-like behavior on Windows.
        * Such that it works in all shells and works with npx.
        */
-      ["", ".cmd", ".ps1"].forEach((ext) => {
-        if (fs.existsSync(nxPath + ext))
-          fs.writeFileSync(ngPath + ext, fs.readFileSync(nxPath + ext));
+      ['', '.cmd', '.ps1'].forEach((ext) => {
+        if (fs.existsSync(nxPath + ext)) fs.writeFileSync(ngPath + ext, fs.readFileSync(nxPath + ext));
       });
     } else {
       // If unix-based, symlink
@@ -75,9 +74,7 @@ function symlinkNgCLItoNxCLI() {
     }
   } catch (e) {
     output.error({
-      title:
-        "Unable to create a symlink from the Angular CLI to the Nx CLI:" +
-        e.message,
+      title: 'Unable to create a symlink from the Angular CLI to the Nx CLI:' + e.message,
     });
     throw e;
   }
@@ -87,10 +84,10 @@ try {
   symlinkNgCLItoNxCLI();
   patchAngularCLI(angularCLIInitPath);
   output.log({
-    title: "Angular CLI has been decorated to enable computation caching.",
+    title: 'Angular CLI has been decorated to enable computation caching.',
   });
 } catch (e) {
   output.error({
-    title: "Decoration of the Angular CLI did not complete successfully",
+    title: 'Decoration of the Angular CLI did not complete successfully',
   });
 }
